+++ src/auth/src/credentials/gdch.rs
@@ -14,11 +14,10 @@
 
 //! [Google Distributed Cloud] service identity authentication.
 
-use crate::Result;
 use crate::credentials::errors::CredentialsError;
 use crate::credentials::service_account::jws::{JwsClaims, JwsHeader};
+use crate::{Result, errors};
 use base64::prelude::{BASE64_URL_SAFE_NO_PAD, Engine as _};
-use rustls::crypto::CryptoProvider;
 use rustls::sign::Signer;
 use rustls_pki_types::PrivateKeyDer;
 use rustls_pki_types::pem::PemObject;
@@ -29,7 +28,6 @@ use serde::Deserialize;
 struct GdchServiceAccountKey {
     /// The credential type, must be "gdch_service_account".
     #[serde(rename = "type")]
-    #[allow(dead_code)]
     cred_type: String,
     /// The format version of the JSON file.
     format_version: String,
@@ -42,7 +40,6 @@ struct GdchServiceAccountKey {
     /// The name of the service identity.
     name: String,
     /// Optional path to custom CA certificate for TLS verification.
-    #[allow(dead_code)]
     ca_cert_path: Option<String>,
     /// The URI to exchange the JWT for a token.
     token_uri: String,
@@ -52,22 +49,13 @@ impl GdchServiceAccountKey {
     #[allow(dead_code)]
     fn signer(&self) -> std::result::Result<Box<dyn Signer>, CredentialsError> {
         let private_key = self.private_key.clone();
-        let key_provider = CryptoProvider::get_default().map(|p| p.key_provider);
-        #[cfg(feature = "default-rustls-provider")]
-        let key_provider = key_provider
-            .unwrap_or_else(|| rustls::crypto::aws_lc_rs::default_provider().key_provider);
-        #[cfg(not(feature = "default-rustls-provider"))]
-        let key_provider = key_provider
-            .expect("The default rustls::CryptoProvider should be configured by the application.");
+        let key_provider = crate::credentials::crypto_provider::get_key_provider();
 
         let key_der = PrivateKeyDer::from_pem_slice(private_key.as_bytes()).map_err(|e| {
-            CredentialsError::from_msg(
-                false,
-                format!(
-                    "Failed to parse GDCH service account private key PEM: {}",
-                    e
-                ),
-            )
+            errors::non_retryable_from_str(format!(
+                "failed to parse GDCH service account private key PEM: {}",
+                e,
+            ))
         })?;
 
         let pk = key_provider
@@ -75,15 +63,13 @@ impl GdchServiceAccountKey {
             .map_err(|e| CredentialsError::from_source(false, e))?;
 
         pk.choose_scheme(&[rustls::SignatureScheme::ECDSA_NISTP256_SHA256])
-            .ok_or_else(|| {
-                CredentialsError::from_msg(
-                    false,
-                    "Unable to choose ECDSA_NISTP256_SHA256 signing scheme as it is not supported by current signer",
-                )
-            })
+            .ok_or_else(|| errors::non_retryable_from_str(
+                "unable to choose ECDSA_NISTP256_SHA256 signing scheme as it is not supported by current signer",
+            ))
     }
 }
 
+// Implements `Debug` for `GdchServiceAccountKey` to avoid printing the private key.
 impl std::fmt::Debug for GdchServiceAccountKey {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("GdchServiceAccountKey")
@@ -149,9 +135,9 @@ impl GdchServiceAccountTokenProvider {
 
         let sig_der = signer
             .sign(to_sign.as_bytes())
-            .map_err(|e| CredentialsError::from_source(false, e))?;
+            .map_err(errors::non_retryable)?;
         let sig = p256::ecdsa::Signature::from_der(&sig_der).map_err(|e| {
-            CredentialsError::from_msg(false, format!("failed to parse ecdsa DER signature: {}", e))
+            errors::non_retryable_from_str(format!("failed to parse ecdsa DER signature: {}", e))
         })?;
         let encoded_sig = BASE64_URL_SAFE_NO_PAD.encode(&sig.to_bytes()[..]);
